+++ server/project-service/src/main/java/edu/illinois/ncsa/incore/service/project/controllers/ProjectController.java
@@ -399,14 +399,18 @@ public class ProjectController {
     public List<DatasetResource> listDatasetsOfProject(
         @Parameter(name = "projectId", description = "ID of the project.") @PathParam("projectId") String id,
         @Parameter(name = "Skip the first n results", description = "Number of results to skip.") @QueryParam("skip") @DefaultValue("0") int offset,
-        @Parameter(name = "Limit the number of results", description = "Maximum number of results to return.") @QueryParam("limit") @DefaultValue("100") int limit,
+        @Parameter(name = "Limit the number of results", description = "Maximum number of results to return. Use -1 to return all results.") @QueryParam("limit") @DefaultValue("100") int limit,
         @Parameter(name = "Filter by type", description = "Filter datasets by type") @QueryParam("type") String type,
+        @Parameter(name = "Filter by workflowId", description = "Filter datasets by workflow id") @QueryParam("workflowId") String workflowId,
+        @Parameter(name = "Filter by executionId", description = "Filter datasets by execution id") @QueryParam("executionId") String executionId,
         @Parameter(name = "Filter by format", description = "Filter datasets by format") @QueryParam("format") String format,
         @Parameter(name = "Text to search ") @QueryParam("text") String text,
         @Parameter(name = "Specify the field or attribute on which the sorting is to be performed.") @DefaultValue("date") @QueryParam("sortBy") String sortBy,
         @Parameter(name = "Specify the order of sorting, either ascending or descending.") @DefaultValue("desc") @QueryParam("order") String order) {
 
         Project project = projectDAO.getProjectById(id);
+        int effectiveLimit = (limit < 0) ? Integer.MAX_VALUE : limit;
+
         if (project != null) {
             if (authorizer.canUserReadMember(username, id, spaceRepository.getAllSpaces(), groups)) {
                 Comparator<DatasetResource> comparator = datasetComparator(sortBy, order);
@@ -414,9 +418,11 @@ public class ProjectController {
                     .filter(dataset -> text == null || dataset.matchesSearchText(text))
                     .filter(dataset -> type == null || dataset.getDataType().equalsIgnoreCase(type))
                     .filter(dataset -> format == null || dataset.format.equalsIgnoreCase(format))
+                    .filter(dataset -> workflowId == null || dataset.hasWorkflowId(dataset, workflowId))
+                    .filter(dataset -> executionId == null || dataset.hasExecutionId(dataset, executionId))
                     .sorted(comparator)
                     .skip(offset)
-                    .limit(limit)
+                    .limit(effectiveLimit)
                     .collect(Collectors.toList());
             } else {
                 throw new IncoreHTTPException(Response.Status.FORBIDDEN, this.username + " does not have privileges to access the project with id " + id);
@@ -508,6 +514,92 @@ public class ProjectController {
             return updatedProject;
         }
         throw new IncoreHTTPException(Response.Status.INTERNAL_SERVER_ERROR, "Failed to delete datasets from the project.");
+    }
+
+    @PATCH
+    @Path("{projectId}/datasets/{datasetId}")
+    @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
+    @Produces(MediaType.APPLICATION_JSON)
+    @Operation(description = "Patch a single dataset within a project")
+    public DatasetResource patchDatasetById(
+        @Parameter(name = "projectId", description = "ID of the project to update")
+        @PathParam("projectId") String projectId,
+        @Parameter(name = "datasetId", description = "ID of the dataset to update")
+        @PathParam("datasetId") String datasetId,
+        @FormParam("title") String title,
+        @FormParam("description") String description,
+        @FormParam("creator") String creator,
+        @FormParam("owner") String owner,
+        @FormParam("format") String format,
+        @FormParam("type") String type,
+        @FormParam("sourceDataset") String sourceDataset,
+        @FormParam("workflowMetadata") String workflowMetadataJson
+    ) {
+        Project project = projectDAO.getProjectById(projectId);
+        if (project == null) {
+            throw new IncoreHTTPException(Response.Status.NOT_FOUND, "Could not find a project with id " + projectId);
+        }
+
+        // Authorization check
+        boolean isAdmin = Authorizer.getInstance().isUserAdmin(this.groups);
+        if (!this.username.equals(project.getOwner()) && !isAdmin) {
+            throw new IncoreHTTPException(Response.Status.FORBIDDEN, this.username + " is not allowed to modify the project.");
+        }
+
+        DatasetResource dataset = project.getDatasets().stream()
+            .filter(d -> d.getId().equals(datasetId))
+            .findFirst()
+            .orElseThrow(() -> new IncoreHTTPException(Response.Status.NOT_FOUND, "Dataset not found: " + datasetId));
+
+        if (title != null) dataset.title = title;
+        if (description != null) dataset.description = description;
+        if (format != null) dataset.format = format;
+        if (type != null) dataset.setDataType(type);
+        if (creator != null) dataset.setDataType(creator);
+        if (owner != null) dataset.setDataType(owner);
+        if (sourceDataset != null) dataset.setDataType(sourceDataset);
+
+        // patch metadata
+        if (workflowMetadataJson != null) {
+            ObjectMapper mapper = new ObjectMapper();
+            try {
+
+                List<WorkflowMetadata> newMetadataList = Arrays.asList(
+                    mapper.readValue(workflowMetadataJson, WorkflowMetadata[].class)
+                );
+
+                for (WorkflowMetadata newMetadata : newMetadataList) {
+                    if (dataset.workflowMetadata == null) {
+                        dataset.workflowMetadata = new ArrayList<>();
+                    }
+
+                    List<WorkflowMetadata> metadataList = dataset.workflowMetadata;
+                    boolean updated = false;
+
+                    for (WorkflowMetadata existing : metadataList) {
+                        if (Objects.equals(existing.getWorkflowId(), newMetadata.getWorkflowId()) &&
+                            Objects.equals(existing.getExecutionId(), newMetadata.getExecutionId())) {
+
+                            WorkflowMetadata.Role newRole =
+                                newMetadata.mergeRoles(existing.getRole(), newMetadata.getRole());
+                            existing.setRole(newRole);
+                            updated = true;
+                            break;
+                        }
+                    }
+
+                    if (!updated) {
+                        metadataList.add(newMetadata);
+                    }
+                }
+
+            } catch (IOException e) {
+                throw new IncoreHTTPException(Response.Status.BAD_REQUEST, "Invalid workflowMetadata input: " + e.getMessage());
+            }
+        }
+
+        projectDAO.updateProject(projectId, project);
+        return dataset;
     }
 
     @GET