+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Endpoint to generate NSI based building inventory dataset by fips code [#370](https://github.com/IN-CORE/incore-services/issues/370)
 - Endpoint to download the generated NSI based building inventory dataset by fips code [#377](https://github.com/IN-CORE/incore-services/issues/377)
 - Endpoint to query NSI based building inventory by bounding box [#379](https://github.com/IN-CORE/incore-services/issues/379)
+- Workflow metadata support, including filtering datasets by workflow ID and execution ID, and a new PATCH endpoint to update individual dataset metadata. [#380](https://github.com/IN-CORE/incore-services/issues/380)
 
 ## [1.28.0] - 2025-02-20
 ### Added